Add files via upload
Note that files can only be posted by admins. 
A relationship between admins and posts is not necessary because it is already taken care of through users. You can add it in, but it is extra maintenance
The files can be considered a BLOB datatype but I used the varchar datatype to represent the url or filepath of the file.